% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{as.symDMatrix.list}
\alias{as.symDMatrix.list}
\title{Coerce a List to a symDMatrix Object.}
\usage{
\method{as.symDMatrix}{list}(x, centers = 0L, scales = 1L, ...)
}
\arguments{
\item{x}{A character vector with names of the \code{ff} files that contain the
data needed to create the object. The files must be ordered by block, \code{G11, G12, G13, ..., G1q, G22, G23, ..., G2q, ..., Gqq}.}

\item{centers}{A numeric vector to fill the \code{@centers} slot of the
\linkS4class{symDMatrix} object.}

\item{scales}{A numeric vector to fill the \code{@scales} slot of the
\linkS4class{symDMatrix} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \linkS4class{symDMatrix} object.
}
\description{
Coerce a list of names of \code{ff} files to a \linkS4class{symDMatrix} object.
}
