% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{blocks}
\alias{blocks}
\title{Returns the block structure of a [symDMatrix-class] object.}
\usage{
blocks(x)
}
\arguments{
\item{x}{A [symDMatrix-class] object.}
}
\description{
Returns the block structure of a [symDMatrix-class] object.
}
