% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{blockSize}
\alias{blockSize}
\title{Returns the Column/Row Block Size of a symDMatrix Object.}
\usage{
blockSize(x)
}
\arguments{
\item{x}{A \linkS4class{symDMatrix} object.}
}
\description{
The last column/row block may be smaller.
}
