% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{chunks}
\alias{chunks}
\title{Returns the chunk structure of a \code{\linkS4class{symDMatrix}} object.}
\usage{
chunks(x)
}
\arguments{
\item{x}{A \code{\linkS4class{symDMatrix}} object.}
}
\description{
Returns the chunk structure of a \code{\linkS4class{symDMatrix}} object.
}

