% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{chunkSize}
\alias{chunkSize}
\title{Returns the column (also row) chunk size of a \code{\linkS4class{symDMatrix}}
object. Note, the last column/row block may be smaller.}
\usage{
chunkSize(x)
}
\arguments{
\item{x}{A \code{\linkS4class{symDMatrix}} object.}
}
\description{
Returns the column (also row) chunk size of a \code{\linkS4class{symDMatrix}}
object. Note, the last column/row block may be smaller.
}

