% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\docType{methods}
\name{diag,symDMatrix-method}
\alias{diag,symDMatrix-method}
\title{Extract the diagonal of a \code{\linkS4class{symDMatrix}}}
\usage{
\S4method{diag}{symDMatrix}(x)
}
\arguments{
\item{x}{a matrix, vector or 1D array, or missing.}
}
\description{
Extract the diagonal of a \code{\linkS4class{symDMatrix}}
}

