% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{nBlocks}
\alias{nBlocks}
\title{Determines the number of column/row blocks of a \linkS4class{symDMatrix} object.}
\usage{
nBlocks(x)
}
\arguments{
\item{x}{A \linkS4class{symDMatrix} object.}
}
\description{
Determines the number of column/row blocks of a \linkS4class{symDMatrix} object.
}
