% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{nChunks}
\alias{nChunks}
\title{Determines the number of column/row chunks of a
\code{\linkS4class{symDMatrix}} object.}
\usage{
nChunks(x)
}
\arguments{
\item{x}{A \code{\linkS4class{symDMatrix}} object.}
}
\description{
Determines the number of column/row chunks of a
\code{\linkS4class{symDMatrix}} object.
}

