% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\docType{class}
\name{symDMatrix-class}
\alias{symDMatrix-class}
\title{A Class to Represent a Symmetric Matrix Paritioned into Memory-Mapped Blocks.}
\description{
A \code{symDMatrix} is a symmetric matrix partitioned into memory-mapped blocks.
Because the matrix is symmetric, only the diagonal and upper-triangular
blocks are stored. Each block is an \code{ff} object.
}
\details{
Internally, the blocks are organized as a nested list in the \code{@data} slot,
each list element representing one row of the symmetric matrix. All blocks
except the ones in the last column/row are expected to have the same
dimensions.
}
