% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{symDMatrix-package}
\alias{symDMatrix-package}
\title{A Package Providing Symmetric Matrices Assembled from Memory-Mapped Blocks.}
\description{
A Package Providing Symmetric Matrices Assembled from Memory-Mapped Blocks.
}
