% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{symDMatrix}
\alias{symDMatrix}
\title{An interface for creating symDMatrix objects}
\usage{
symDMatrix(dataFiles, centers = 0, scales = 1, names = character())
}
\description{
An interface for creating symDMatrix objects
}

