% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{as.symDMatrix.matrix}
\alias{as.symDMatrix.matrix}
\title{Coerce a Matrix to a symDMatrix Object.}
\usage{
\method{as.symDMatrix}{matrix}(x, blockSize = 5000L, vmode = "double",
  folderOut = randomString(), ...)
}
\arguments{
\item{x}{A numeric matrix (assumed to be symmetric).}

\item{blockSize}{The number of rows and columns of each block. If \code{NULL}, a
single block of the same dimensions as \code{x} will be created. Defaults to
5000.}

\item{vmode}{The vmode used to store the data in the \code{ff} objects.}

\item{folderOut}{A name for a folder where to store the data of the
resulting \linkS4class{symDMatrix} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \linkS4class{symDMatrix} object.
}
\description{
Coerce a numeric matrix (assumed to be symmetric) in RAM to a
\linkS4class{symDMatrix} object. Saves the metadata to reload the matrix using
\code{\link[=load.symDMatrix]{load.symDMatrix()}} as \code{symDMatrix.RData}.
}
