% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{as.symDMatrix.matrix}
\alias{as.symDMatrix.matrix}
\title{Coerce a Matrix to a symDMatrix Object.}
\usage{
\method{as.symDMatrix}{matrix}(x, nBlocks = 3L, vmode = "double",
  folder = randomString(), saveRData = TRUE, ...)
}
\arguments{
\item{x}{A numeric matrix (assumed to be symmetric).}

\item{nBlocks}{The number of column (also row) blocks to be used.}

\item{vmode}{The vmode used to store the data in the \code{ff} objects.}

\item{folder}{A name for a folder where to store the data of the resulting
\linkS4class{symDMatrix} object.}

\item{saveRData}{If TRUE, the metadata (the \linkS4class{symDMatrix}) is saved
using the name \code{G.RData}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \linkS4class{symDMatrix} object.
}
\description{
Coerce a numeric matrix (assumed to be symmetric) in RAM to a
\linkS4class{symDMatrix} object.
}
