% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{as.symDMatrix}
\alias{as.symDMatrix}
\title{Coerce an Object to a symDMatrix Object.}
\usage{
as.symDMatrix(x, ...)
}
\arguments{
\item{x}{A numeric matrix.}

\item{...}{Additional arguments.}
}
\value{
A \linkS4class{symDMatrix} object.
}
\description{
Coerce an object to a \linkS4class{symDMatrix} object.
}
\seealso{
\code{\link[=as.symDMatrix.matrix]{as.symDMatrix.matrix()}} to coerce a matrix or
\code{\link[=as.symDMatrix.character]{as.symDMatrix.character()}} to coerce a vector of path names to a
\linkS4class{symDMatrix} object.
}
