% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{as.symDMatrix}
\alias{as.symDMatrix}
\alias{as.symDMatrix.matrix}
\title{Coerce an Object to a symDMatrix Object.}
\usage{
as.symDMatrix(x, ...)

\method{as.symDMatrix}{matrix}(x, nBlocks = 3L, vmode = "double",
  folder = randomString(), saveRData = TRUE)
}
\arguments{
\item{x}{A numeric matrix.}

\item{...}{Additional arguments.}
}
\value{
A \linkS4class{symDMatrix} object.
}
\description{
Coerce an object to a \linkS4class{symDMatrix} object.
}
