\name{as.symDMatrix}
\alias{as.symDMatrix}
\title{Coerce an Object to a symDMatrix Object}
\description{
    Coerce an object to a \code{symDMatrix} object.
}
\usage{
as.symDMatrix(x, ...)
}
\arguments{
    \item{x}{
        A numeric matrix.
    }
    \item{...}{
        Additional arguments.
    }
}
\value{
    A \code{symDMatrix} object.
}
\seealso{
    \code{\link{as.symDMatrix.matrix}} to coerce a matrix or
    \code{\link{as.symDMatrix.character}} to coerce a vector of path names to a
    \code{symDMatrix} object.
}
