% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{blockIndex}
\alias{blockIndex}
\title{Returns the Block Structure of a symDMatrix Object.}
\usage{
blockIndex(x)
}
\arguments{
\item{x}{A \linkS4class{symDMatrix} object.}
}
\value{
A matrix.
}
\description{
Returns the Block Structure of a symDMatrix Object.
}
