% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{blocks}
\alias{blocks}
\title{Returns the Block Structure of a symDMatrix Object.}
\usage{
blocks(x)
}
\arguments{
\item{x}{A \linkS4class{symDMatrix} object.}
}
\description{
Returns the Block Structure of a symDMatrix Object.
}
