% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{blockSize}
\alias{blockSize}
\title{Returns the column/row block size of a [symDMatrix-class] object. Note, the
last column/row block may be smaller.}
\usage{
blockSize(x)
}
\arguments{
\item{x}{A [symDMatrix-class] object.}
}
\description{
Returns the column/row block size of a [symDMatrix-class] object. Note, the
last column/row block may be smaller.
}
