% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\docType{methods}
\name{initialize,symDMatrix-method}
\alias{initialize,symDMatrix-method}
\title{Creates a New symDMatrix Instance.}
\usage{
\S4method{initialize}{symDMatrix}(.Object, data, centers = 0L, scales = 1L)
}
\arguments{
\item{.Object}{The \linkS4class{symDMatrix} instance to be initialized. This
argument is passed in by R and can be ignored, but still needs to be
documented.}

\item{data}{A nested list in the form of \code{[[G11, G12, G13, ..., G1q],
[G22, G23, ..., G2q], [...], [Gqq]]} containing the blocks of the
upper-triangular symmetric matrix.}

\item{centers}{A numeric vector storing the values used for column centering
when creating the symmetric matrix.}

\item{scales}{A numeric vector storing the values used for column scaling
when creating the symmetric matrix.}
}
\description{
This method is run when a \linkS4class{symDMatrix} object is created using
\code{symDMatrix(...)} or \code{new("symDMatrix", ...)}.
}
