% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\docType{methods}
\name{initialize,symDMatrix-method}
\alias{initialize,symDMatrix-method}
\title{Creates a New symDMatrix Instance.}
\usage{
\S4method{initialize}{symDMatrix}(.Object, data, centers = 0L, scales = 1L)
}
\arguments{
\item{.Object}{The \linkS4class{symDMatrix} instance to be initialized. This
argument is passed in by R and can be ignored, but still needs to be
documented.}

\item{data}{A nested list to fill the \code{@data} slot of the
\linkS4class{symDMatrix} object.}

\item{centers}{A numeric vector to fill the \code{@centers} slot of the
\linkS4class{symDMatrix} object.}

\item{scales}{A numeric vector to fill the \code{@scales} slot of the
\linkS4class{symDMatrix} object.}
}
\description{
This method is run when a \linkS4class{symDMatrix} object is created using
\code{symDMatrix(...)} or \code{new("symDMatrix", ...)}.
}
