% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{load.symDMatrix}
\alias{load.symDMatrix}
\title{Loads symDMatrix Objects from .RData Files.}
\usage{
load.symDMatrix(file, envir = parent.frame())
}
\arguments{
\item{file}{The name of an .RData file to be loaded.}

\item{envir}{The environment where to load the data.}
}
\description{
This function is similar to \code{\link[base:load]{base::load()}}, but also initializes the
different matrix-like objects that \linkS4class{symDMatrix} can take. Currently
supported are \code{ff_matrix} objects.
}
