% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{nBlocks}
\alias{nBlocks}
\title{Return the Number of Column/Row Blocks of a symDMatrix Object.}
\usage{
nBlocks(x)
}
\arguments{
\item{x}{A \linkS4class{symDMatrix} object.}
}
\value{
The number of column/row blocks of a \linkS4class{symDMatrix} object.
}
\description{
This function returns the number of row blocks the original matrix has been partitioned into.
}
\examples{
# Load example symDMatrix (G)
load.symDMatrix(system.file("extdata", "G.RData", package = "symDMatrix"))

# Get the number of row blocks the original matrix was partitioned into
nBlocks(G)
}
