% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\docType{class}
\name{symDMatrix-class}
\alias{symDMatrix-class}
\title{An S4 class that represents a memory-mapped symmetric matrix.}
\description{
An S4 class that represents a memory-mapped symmetric matrix.
}

