% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\docType{class}
\name{symDMatrix-class}
\alias{symDMatrix-class}
\title{An S4 class that represents a symmetric matrix assembled from memory-mapped
blocks.}
\description{
A [symDMatrix-class] is a symmetric matrix assembled from memory-mapped
blocks. Because the matrix is symmetric, only the diagonal and
upper-triangular blocks are stored. Each block is an `ff` object.
}
\details{
Internally, the blocks are organized as a nested list in the `@data` slot,
each list element representing one row of the symmetric matrix.
}
