% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symDMatrix.R
\name{symDMatrix}
\alias{symDMatrix}
\title{A constructor for creating \code{\linkS4class{symDMatrix}} objects.}
\usage{
symDMatrix(dataFiles, centers = 0, scales = 1)
}
\arguments{
\item{dataFiles}{A character vector with names of the \code{ff} files that
contain the data needed to create the object. The files must be ordered by
block, G11, G12, G13, ..., G1q, G22, G23, ..., Gqq.}

\item{centers}{(numeric) A vector storing the means used, if any, when
creating the symmetric matrix.}

\item{scales}{(numeric) A vector storing the standard deviations used, if
any, when creating the symmetric matrix.}
}
\value{
A \code{\linkS4class{symDMatrix}} object.
}
\description{
A \code{\linkS4class{symDMatrix}} is a symmetric matrix assembled from
memory-mapped blocks. Because the matrix is symmetric, only the diagonal and
upper-triangular blocks are stored. Each block is an \code{ff} object.
}

